+++ src/ServiceProvider.php
@@ -13,7 +13,7 @@ use Statamic\Providers\AddonServiceProvider;
 
 class ServiceProvider extends AddonServiceProvider
 {
-    const VERSION = '1.2.10';
+    const VERSION = '1.2.11';
 
     protected $tags = [
         Kai::class,